# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PelicanTownSpecials API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/assets/images": {
      "post": {
        "summary": "Upload Image",
        "operationId": "upload_image_api_v1_assets_images_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image_api_v1_assets_images_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{asset_id}": {
      "get": {
        "summary": "Get Image",
        "operationId": "get_image_api_v1_assets__asset_id__get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drafts": {
      "get": {
        "summary": "List Drafts",
        "operationId": "list_drafts_api_v1_drafts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_DraftSummary_"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Draft",
        "operationId": "create_draft_api_v1_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drafts/{draft_id}": {
      "get": {
        "summary": "Get Draft",
        "operationId": "get_draft_api_v1_drafts__draft_id__get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch Draft",
        "operationId": "patch_draft_api_v1_drafts__draft_id__patch",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drafts/{draft_id}/convert-to-blueprint": {
      "post": {
        "summary": "Convert To Blueprint",
        "operationId": "convert_to_blueprint_api_v1_drafts__draft_id__convert_to_blueprint_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drafts/{draft_id}/archive": {
      "post": {
        "summary": "Archive Draft",
        "operationId": "archive_draft_api_v1_drafts__draft_id__archive_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CookbookDishDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drafts/{draft_id}/discard": {
      "post": {
        "summary": "Discard Draft",
        "operationId": "discard_draft_api_v1_drafts__draft_id__discard_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cookbook/{dish_id}": {
      "get": {
        "summary": "Get Cookbook Dish",
        "operationId": "get_cookbook_dish_api_v1_cookbook__dish_id__get",
        "parameters": [
          {
            "name": "dish_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dish Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CookbookDishDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Cookbook Dish",
        "operationId": "delete_cookbook_dish_api_v1_cookbook__dish_id__delete",
        "parameters": [
          {
            "name": "dish_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dish Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssetKind": {
        "type": "string",
        "enum": [
          "ORIGINAL_IMAGE",
          "GENERATED_ART",
          "PREVIEW",
          "ICON_SOURCE",
          "ICON_16",
          "MOD_SPRITESHEET",
          "EXPORT_ZIP"
        ],
        "title": "AssetKind"
      },
      "AssetView": {
        "properties": {
          "assetId": {
            "type": "string",
            "format": "uuid",
            "title": "Assetid"
          },
          "kind": {
            "$ref": "#/components/schemas/AssetKind"
          },
          "mediaType": {
            "$ref": "#/components/schemas/MediaType"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "byteSize": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Bytesize"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 8192.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 8192.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "sourceRevision": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcerevision"
          },
          "attemptId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attemptid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "assetId",
          "kind",
          "mediaType",
          "sha256",
          "byteSize",
          "createdAt"
        ],
        "title": "AssetView"
      },
      "Body_upload_image_api_v1_assets_images_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Original dish photo"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_image_api_v1_assets_images_post"
      },
...
      "CookbookDishDetail": {
        "properties": {
          "dishId": {
            "type": "string",
            "format": "uuid",
            "title": "Dishid"
          },
          "archivedAt": {
            "type": "string",
            "title": "Archivedat"
          },
          "displayName": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Displayname"
          },
          "internalName": {
            "type": "string",
            "maxLength": 48,
            "minLength": 3,
            "title": "Internalname"
          },
          "categoryLabel": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Categorylabel"
          },
          "description": {
            "type": "string",
            "maxLength": 400,
            "minLength": 1,
            "title": "Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Tags"
          },
          "gameplay": {
            "$ref": "#/components/schemas/GameplaySpec"
          },
          "visuals": {
            "$ref": "#/components/schemas/CookbookVisuals"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dishId",
          "archivedAt",
          "displayName",
          "internalName",
          "categoryLabel",
          "description",
          "gameplay",
          "visuals"
        ],
        "title": "CookbookDishDetail"
      },
...
      "DishAnalysis": {
        "properties": {
          "recognizedDish": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Recognizeddish"
          },
          "summary": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Summary"
          },
          "cuisine": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Cuisine"
          },
          "cookingMethods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "title": "Cookingmethods"
          },
          "flavorProfile": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Flavorprofile"
          },
          "semanticIngredients": {
            "items": {
              "$ref": "#/components/schemas/SemanticIngredient"
            },
            "type": "array",
            "maxItems": 12,
            "minItems": 1,
            "title": "Semanticingredients"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "safetyNotes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Safetynotes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "recognizedDish",
          "summary",
          "semanticIngredients",
          "confidence"
        ],
        "title": "DishAnalysis"
      },
      "DraftCreateRequest": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/DraftMode"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "source": {
            "$ref": "#/components/schemas/DraftCreateSource"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "language",
          "source"
        ],
        "title": "DraftCreateRequest"
      },
      "DraftCreateSource": {
        "properties": {
          "originalImageAssetId": {
            "type": "string",
            "format": "uuid",
            "title": "Originalimageassetid"
          },
          "contextText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contexttext"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "originalImageAssetId"
        ],
        "title": "DraftCreateSource"
      },
      "DraftMode": {
        "type": "string",
        "enum": [
          "ASK_GUS",
          "BLUEPRINT"
        ],
        "title": "DraftMode"
      },
      "DraftPatchRequest": {
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expectedrevision"
          },
          "presentation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PresentationSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "gameplay": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GameplaySpec"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expectedRevision"
        ],
        "title": "DraftPatchRequest"
      },
      "DraftStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "READY",
          "GENERATING",
          "REGENERATING",
          "REVIEWABLE",
          "STALE_PREVIEW",
          "FAILED",
          "ARCHIVED",
          "DISCARDED"
        ],
        "title": "DraftStatus"
      },
      "DraftSummary": {
        "properties": {
          "draftId": {
            "type": "string",
            "format": "uuid",
            "title": "Draftid"
          },
          "mode": {
            "$ref": "#/components/schemas/DraftMode"
          },
          "status": {
            "$ref": "#/components/schemas/DraftStatus"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "displayName": {
            "type": "string",
            "maxLength": 60,
            "title": "Displayname"
          },
          "originalImageAssetId": {
            "type": "string",
            "format": "uuid",
            "title": "Originalimageassetid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "draftId",
          "mode",
          "status",
          "revision",
          "updatedAt",
          "displayName",
          "originalImageAssetId"
        ],
        "title": "DraftSummary"
      },
      "DraftView": {
        "properties": {
          "draftId": {
            "type": "string",
            "format": "uuid",
            "title": "Draftid"
          },
          "mode": {
            "$ref": "#/components/schemas/DraftMode"
          },
          "baseTemplateVersion": {
            "anyOf": [
              {
                "type": "string",
                "const": "blueprint-v1"
              },
              {
                "type": "null"
              }
            ],
            "title": "Basetemplateversion"
          },
          "status": {
            "$ref": "#/components/schemas/DraftStatus"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "source": {
            "$ref": "#/components/schemas/SourceInput"
          },
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DishAnalysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "presentation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PresentationSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "gameplay": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GameplaySpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "visuals": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VisualSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "lastError": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "archivedDishId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archiveddishid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "draftId",
          "mode",
          "baseTemplateVersion",
          "status",
          "revision",
          "source",
          "provenance",
          "createdAt",
          "updatedAt",
          "archivedDishId"
        ],
        "title": "DraftView"
      },
...
      "GameplaySpec": {
        "properties": {
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/GameIngredient"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Ingredients"
          },
          "recovery": {
            "$ref": "#/components/schemas/RecoverySpec"
          },
          "buff": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BuffSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "sellPrice": {
            "type": "integer",
            "maximum": 50000.0,
            "minimum": 0.0,
            "title": "Sellprice"
          },
          "isDrink": {
            "type": "boolean",
            "title": "Isdrink"
          },
          "recipeUnlock": {
            "$ref": "#/components/schemas/RecipeUnlock",
            "default": "DEFAULT"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ingredients",
          "recovery",
          "sellPrice",
          "isDrink"
        ],
        "title": "GameplaySpec"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Language": {
        "type": "string",
        "enum": [
          "zh-CN",
          "en-US"
        ],
        "title": "Language"
      },
...
      "MediaType": {
        "type": "string",
        "enum": [
          "image/png",
          "image/jpeg",
          "image/webp",
          "application/zip"
        ],
        "title": "MediaType"
      },
...
      "Page_DraftSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DraftSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "Page[DraftSummary]"
      },
      "PresentationSpec": {
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Displayname"
          },
          "internalName": {
            "type": "string",
            "maxLength": 48,
            "minLength": 3,
            "pattern": "^[A-Za-z][A-Za-z0-9_]{2,47}$",
            "title": "Internalname"
          },
          "categoryLabel": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Categorylabel"
          },
          "description": {
            "type": "string",
            "maxLength": 400,
            "minLength": 1,
            "title": "Description"
          },
          "gusComment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 400
              },
              {
                "type": "null"
              }
            ],
            "title": "Guscomment"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 12,
            "minItems": 0,
            "title": "Tags"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "displayName",
          "internalName",
          "categoryLabel",
          "description"
        ],
        "title": "PresentationSpec"
      },
...
      "SemanticIngredient": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "normalizedName": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Normalizedname"
          },
          "visibleConfidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Visibleconfidence"
          },
          "quantityHint": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantityhint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "normalizedName",
          "visibleConfidence"
        ],
        "title": "SemanticIngredient"
      },
      "SourceInput": {
        "properties": {
          "originalImageAssetId": {
            "type": "string",
            "format": "uuid",
            "title": "Originalimageassetid"
          },
          "contextText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contexttext"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "originalImageAssetId",
          "language"
        ],
        "title": "SourceInput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}